% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOQ_bias_precision_functions.R
\name{calc_bias}
\alias{calc_bias}
\title{Function for calculating LOQ}
\usage{
calc_bias(MRD, input_cancer_cells, bias.plot = TRUE, log = TRUE,
  log_offset = 1e-06)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{input_cancer_cells}{The pre-specified input.  Note: can be frequency or any other input value}

\item{bias.plot}{If TRUE, returns a ggplot of the bias}

\item{log}{Boolean for log10-transforming input_cancer_cells, default TRUE}

\item{log_offset}{log offset for transformation to avoid log(0), default 1e-6}
}
\value{
Returns a list containing results: a data.frame with bias and relative bias.  If bias.plot == TRUE, also contains a ggplot of bias vs. input cancer cells.
}
\description{
Function for calculating LOQ
}
