% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOQ_bias_precision_functions.R
\name{calc_precision}
\alias{calc_precision}
\title{Function for calculating MRD precision}
\usage{
calc_precision(MRD, input_cancer_cells, precision.plot = TRUE, log = TRUE,
  log_offset = 1e-06, precision_type = "CV", sadler_params = list(b1 = 2,
  b2 = 0.1, b3 = -1.8))
}
\arguments{
\item{MRD}{The measured MRD values}

\item{input_cancer_cells}{The pre-specified input.  Note: can be frequency or any other input value}

\item{precision.plot}{If TRUE, returns a ggplot of the Precision and Sadler's fit}

\item{log}{Boolean for log10-transforming input_cancer_cells, default TRUE}

\item{log_offset}{log offset for transformation to avoid log(0), default 1e-6}

\item{precision_type}{The type of precision.  Either percent CV ("CV") or relative S.D. ("rel_SD") , default "CV"}

\item{sadler_params}{A list of initial values for sadler's fit}
}
\value{
Returns a list containing LOQ, LCL.LOQ and UCL.LOQ.  Note that LCL.LOQ will almost always be NA; use LOD as the lower limit. If TE.plot == TRUE, also contains TE_fit, fitted_data and TE_plot.
}
\description{
Function for calculating MRD precision
}
