% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_model_assumptions.R
\name{CoxPH_Diagnostics}
\alias{CoxPH_Diagnostics}
\title{Cox-PH model assumptions
This function outputs a bunch of plots related to Cox-PH diagnostics and model assumptions.  Based on "Amendment 2 DFCI Analyses 3November2017.R"}
\usage{
CoxPH_Diagnostics(model, data, continuous_vars, save_prefix)
}
\arguments{
\item{model}{A cox-ph fitted model object}

\item{data}{The data.frame containing the covariates used to generate the model}

\item{continuous_vars}{A character vector of the names of the continuous variables used in the model}

\item{save_prefix}{The file prefix for saving the output.  E.g., "Results/coxph_"}
}
\value{
This function doesn't return anything, but writes four plots to disk: the scaled schoenfeld residuals, dfBetas, Deviance and Martingale residuals.  These plots can be used to e.g. diagnose issues with the model, find outliers or find time-varying trends.
}
\description{
Cox-PH model assumptions
This function outputs a bunch of plots related to Cox-PH diagnostics and model assumptions.  Based on "Amendment 2 DFCI Analyses 3November2017.R"
}
\examples{
data(veteran)
model <- coxph(Surv(time, status) ~ ., data=veteran)
cvars <- c("karno", "diagtime", "age")
CoxPH_Diagnostics(model, data=veteran, continuous_vars=cvars, save_prefix="~/Desktop/coxph_veteran_")
}
\seealso{
"Checking the Cox model with cumulative sums of martingale-based residuals." Biometrika 80.3 (1993): 557-572.
"Using Time Dependent Covariates..." (https://cran.r-project.org/web/packages/survival/vignettes/timedep.pdf)
"Proportional Hazards Regression Diagnostics" (http://www.ics.uci.edu/~dgillen/STAT255/Handouts/lecture10.pdf)
}
