% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{methods}
\name{createKeywordTrees,EOS_Report-method}
\alias{createKeywordTrees,EOS_Report-method}
\title{createKeywordTrees}
\usage{
\S4method{createKeywordTrees}{EOS_Report}(obj)
}
\description{
createKeywordTrees
}
