% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{createKeywordTrees}
\alias{createKeywordTrees}
\title{Function for creating keyword trees
Create keyword trees used for finding compatible sequences between reports.  Call getLocus and spltiSequenceIntoSubstrings first.}
\usage{
createKeywordTrees(obj)
}
\arguments{
\item{obj}{The EOS report object}
}
\value{
The object with left_hash and right_hash slots filled.
}
\description{
Function for creating keyword trees
Create keyword trees used for finding compatible sequences between reports.  Call getLocus and spltiSequenceIntoSubstrings first.
}
