% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{methods}
\name{determineCdr3Index,EOS_Report-method}
\alias{determineCdr3Index,EOS_Report-method}
\title{determineCdr3Index}
\usage{
\S4method{determineCdr3Index}{EOS_Report}(obj)
}
\description{
determineCdr3Index
}
