% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{determineCdr3Index}
\alias{determineCdr3Index}
\title{Function for calculating cdr3 index}
\usage{
determineCdr3Index(obj)
}
\arguments{
\item{obj}{The EOS report object}
}
\value{
The object with cdr3Index slot filled
}
\description{
Function for calculating cdr3 index
}
