% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{methods}
\name{determineNdnValues,EOS_Report-method}
\alias{determineNdnValues,EOS_Report-method}
\title{determineNdnValues}
\usage{
\S4method{determineNdnValues}{EOS_Report}(obj)
}
\description{
determineNdnValues
}
