% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{determineNdnValues}
\alias{determineNdnValues}
\title{Function for calculating NDN base pairs}
\usage{
determineNdnValues(obj)
}
\arguments{
\item{obj}{The EOS report object}
}
\value{
The object with ndnIndex and ndnLength slots filled
}
\description{
Function for calculating NDN base pairs
}
