% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{class}
\name{EOS_Report-class}
\alias{EOS_Report-class}
\title{Object for calibrating and matching sequences}
\description{
Object for calibrating and matching sequences
}
\section{Slots}{

\describe{
\item{\code{trimmed_sequence}}{trimmed sequence}

\item{\code{locus}}{locus determined through getLocus function}

\item{\code{cdr3Index}}{calculated cdr3 index}

\item{\code{cdr3Length}}{calculated cdr3 length}

\item{\code{ndnIndex}}{calculated ndn index}

\item{\code{ndnLength}}{calculated ndn length}

\item{\code{left_sequence}}{left split sequence based on cdr3 index}

\item{\code{right_sequence}}{right split sequence based on cdr3 index}

\item{\code{left_ndn_start}}{left ndn base start}

\item{\code{left_ndn_end}}{left ndn base end}

\item{\code{right_ndn_start}}{left ndn base start}

\item{\code{right_ndn_end}}{left ndn base end}

\item{\code{left_hash}}{tree hash of left sequence}

\item{\code{right_hash}}{tree hash of right sequence}

\item{\code{data}}{The original CSV report data}
}}

\examples{
new("EOS_Report", data=Datain)
}
\seealso{
See examples/PRO-00243_calibration_example.r
}
