% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearity_functions.R
\name{find_linearity}
\alias{find_linearity}
\title{Function for finding linear MRD range}
\usage{
find_linearity(MRD, input_cancer_cells, max_deviation_threshold = 0.05,
  log = TRUE, log_offset = 1e-06, linearity.plot = TRUE,
  log_deviation = T)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{input_cancer_cells}{The pre-specified input.  Note: can be frequency or any other input value}

\item{log}{Boolean for log10-transforming input_cancer_cells, default TRUE}

\item{log_offset}{log offset for transformation to avoid log(0), default 1e-6}

\item{linearity.plot}{If TRUE, returns a ggplot of the Total Error fit}

\item{log_deviation}{If TRUE and if log==T, calculate deviation in log space.  Otherwise, convert deviance to linear space.}

\item{max_deviance_threshold}{Maximum deviation threshold from linearity}
}
\value{
Returns a list containing linear_range_min, linear_range_max and deviation within the linear range.  If linearity.plot == TRUE, also contains a ggplot of the linear range.
}
\description{
Function for finding linear MRD range
}
\seealso{
PRO-00091
}
