% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOD_function.R
\name{find_LOD}
\alias{find_LOD}
\title{Function for calculating LOD}
\usage{
find_LOD(MRD, input_cancer_cells, prob = 0.95, conf.level = 0.95,
  log = TRUE, log_offset = 1e-06, probit.plot = TRUE)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{input_cancer_cells}{The pre-specified input.  Note: can be frequency or any other input value}

\item{prob}{Detection probability for LOD, default 0.95}

\item{conf.level}{Confidence level for calculating upper and lower bounds, default 0.95}

\item{log}{Boolean for log10-transforming MRD and input_cancer_cells, default TRUE}

\item{log_offset}{If log transforming, offset value for the MRD to avoid log(0) issues, default 0}

\item{probit.plot}{If TRUE, returns a ggplot of the probit fit}
}
\value{
Returns a list containing LOD, LCL.LOD and UCL.LOD.  If probit.plot == TRUE, also contains probit_fit, fitted_data and probit_plot.
}
\description{
Function for calculating LOD
}
\examples{
Datain <- read.table("processed_data/20171119_MVP-00147_lod_data-table1.tsv", head=TRUE, as.is=TRUE, sep="\\t")
find_LOD(Datain$Sample.MRD.Frequency, Datain$Input.Cancer.Cells, probit.plot=F)$LOD
[1] 2.289792
}
