% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOQ_bias_precision_functions.R
\name{find_LOQ}
\alias{find_LOQ}
\title{Function for calculating LOQ}
\usage{
find_LOQ(MRD, input_cancer_cells, threshold = 0.7, conf.level = 0.95,
  log = TRUE, log_offset = 1e-06, TE.plot = TRUE,
  sadler_params = list(b1 = 2, b2 = 0.1, b3 = -1.8), extra_sadler_log = F)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{input_cancer_cells}{The pre-specified input.  Note: can be frequency or any other input value}

\item{conf.level}{Confidence level for calculating upper and lower bounds, default 0.95}

\item{log}{Boolean for log10-transforming input_cancer_cells, default TRUE}

\item{log_offset}{log offset for transformation to avoid log(0), default 1e-6}

\item{TE.plot}{If TRUE, returns a ggplot of the Total Error fit}

\item{sadler_params}{A list of initial values for sadler's fit}

\item{extra_sadler_log}{Whether to take an extra log to improve Sadler fit stability; i.e.: log(y) ~ log(b1+b2*x)^b3)}

\item{prob}{Detection probability for LOD, default 0.95}
}
\value{
Returns a list containing LOQ, LCL.LOQ and UCL.LOQ.  Note that LCL.LOQ will almost always be NA; use LOD as the lower limit. If TE.plot == TRUE, also contains TE_fit, fitted_data and TE_plot.
}
\description{
Function for calculating LOQ
}
