% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{findConsensusSequences}
\alias{findConsensusSequences}
\title{Find consensus sequences
Find consensus sequences within EOS report(s)}
\usage{
findConsensusSequences(..., report_list = NULL, no_shm_allowance = T,
  max_adj_mut_igkl = 0, max_adj_mut_igh = 2, ndn_cost = 1.05)
}
\arguments{
\item{...}{Any number of EOS report objects}

\item{report_list}{A list of EOS report objects}

\item{no_shm_allowance}{Should mutations be allowed?}

\item{max_adj_mut_igkl}{If the EOS report query doesn't have a pre-specified max adjusted mutation value for a sequence, use this default for IGKL sequences}

\item{max_adj_mut_igh}{If the EOS report query doesn't have a pre-specified max adjusted mutation value for a sequence, use this default for IGH/IGH_D sequences}

\item{ndn_cost}{Modified cost for an ndn base pair mismatch}
}
\value{
A list with S3 class "Consensus_List".  Each element in the list is a vector of matching sequences and corresponding locus in the "locus" attribute field.  The names of the list correspond to the dominant sequence of the consensus cluster (i.e., the shortest).
}
\description{
Find consensus sequences
Find consensus sequences within EOS report(s)
}
