% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getLocus}
\alias{getLocus}
\title{Function for calculating gene locus}
\usage{
getLocus(obj)
}
\arguments{
\item{obj}{The EOS report object}
}
\value{
The object with locus slot filled
}
\description{
Function for calculating gene locus
}
