% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getReportMetadata}
\alias{getReportMetadata}
\title{Extract metadata from report
Extract metadata from the header of a report file}
\usage{
getReportMetadata(file_name, directory = "",
  property = "estTotalNucleatedCells")
}
\arguments{
\item{file_name}{The name of the EOS report file.  Can be gzipped; will be determined through file extension.}

\item{directory}{Optional directory location of file}

\item{property}{The name of the property to extract}
}
\value{
The metadata value.  Note: the return will be a string value.
}
\description{
Extract metadata from report
Extract metadata from the header of a report file
}
