% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{getReportSeqsByTag}
\alias{getReportSeqsByTag}
\title{Read in data and select by tag
Helper function for reading in data.  Given a report file name, reads in the data and selects rows based on the sequenceTags column.}
\usage{
getReportSeqsByTag(file_names, directory = "", tag = "Dx")
}
\arguments{
\item{file_names}{A vector of file names}

\item{directory}{Optional directory location of file}

\item{tag}{The tag to select (by exact string matching the field); if `all` return all data.}
}
\value{
A data.table of the report.  If more than one file was named, the reports are combined.
}
\description{
Read in data and select by tag
Helper function for reading in data.  Given a report file name, reads in the data and selects rows based on the sequenceTags column.
}
