% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{matchQuerySequenceWithTargetSequence}
\alias{matchQuerySequenceWithTargetSequence}
\title{Match sequences between objects
Match sequences in one object to sequences in another object using keyword hash trees}
\usage{
matchQuerySequenceWithTargetSequence(obj_query, obj_target,
  no_shm_allowance = T, max_adj_mut_igkl = 0, max_adj_mut_igh = 2,
  ndn_cost = 1.05, include_report_data = F)
}
\arguments{
\item{obj_query}{The query EOS report object}

\item{obj_target}{The target EOS report object (i.e., the one with the trees)}

\item{no_shm_allowance}{Should mutations be allowed?}

\item{max_adj_mut_igkl}{If the EOS report query doesn't have a pre-specified max adjusted mutation value for a sequence, use this default for IGKL sequences}

\item{max_adj_mut_igh}{If the EOS report query doesn't have a pre-specified max adjusted mutation value for a sequence, use this default for IGH/IGH_D sequences}

\item{ndn_cost}{Modified cost for an ndn base pair mismatch}

\item{include_report_data}{If TRUE, append all EOS fields to the data.frame of matched sequences}
}
\value{
A data.frame of matching sequences, locus, superlocus, total mutations and optionally the EOS report fields
}
\description{
Match sequences between objects
Match sequences in one object to sequences in another object using keyword hash trees
}
