% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{matchReportToConsensus}
\alias{matchReportToConsensus}
\title{Match Sequences to a Consensus
Find exact matching sequences in an EOS report object that match to any sequences in a "Consensus_List"}
\usage{
matchReportToConsensus(obj_query, consensus, include_report_data = F)
}
\arguments{
\item{obj_query}{The EOS report query}

\item{consensus}{The Consensus List}

\item{include_report_data}{If TRUE, append all EOS fields to the data.frame of matched sequences}
}
\value{
A data.frame of matching sequence, consensus sequence locus and optionally the EOS report fields
}
\description{
Match Sequences to a Consensus
Find exact matching sequences in an EOS report object that match to any sequences in a "Consensus_List"
}
