% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{processReport}
\alias{processReport}
\title{processReport
Calls functions in the following order: getLocus, determineNdnValues, determineCdr3Index, trimCloneSequence, splitSequenceIntoSubstrings, createKeywordTrees}
\usage{
processReport(obj, seq_edge_trim = 2L)
}
\arguments{
\item{seq_edge_trim}{Value passed to trimCloneSequence}
}
\value{
The processed EOS report object
}
\description{
processReport
Calls functions in the following order: getLocus, determineNdnValues, determineCdr3Index, trimCloneSequence, splitSequenceIntoSubstrings, createKeywordTrees
}
