% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{SICS_model}
\alias{SICS_model}
\title{Separate Intercepts Common Slope
Separate Intercepts Common Slopes for stability analysis}
\usage{
SICS_model(MRD, Time, Sample.ID, UL = 1.3, LL = 0.7,
  stability.plot = TRUE)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{Time}{Time (numeric)}

\item{Sample.ID}{Sample ID used for model selection}

\item{UL}{Upper limit of MRD (in MRD raw scale)}

\item{LL}{Lower limit of MRD (in MRD raw scale)}

\item{stability.plot}{If TRUE, returns a ggplot of the data}
}
\value{
Returns a list the estimated shelf_life and a plot if specified.
}
\description{
Separate Intercepts Common Slope
Separate Intercepts Common Slopes for stability analysis
}
