% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{methods}
\name{splitSequenceIntoSubstrings,EOS_Report-method}
\alias{splitSequenceIntoSubstrings,EOS_Report-method}
\title{splitSequenceIntoSubstrings}
\usage{
\S4method{splitSequenceIntoSubstrings}{EOS_Report}(obj)
}
\description{
splitSequenceIntoSubstrings
}
