% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{splitSequenceIntoSubstrings}
\alias{splitSequenceIntoSubstrings}
\title{Function for splitting sequence into subsequences
This function splits sequences into left and right sequences based on cdr3Index.  Must have cdr3Index, ndnIndex, ndnLength and trimmed_sequence slots filled.  Call determineNdnValues, determineCdr3Index and trimCloneSequence functions first.}
\usage{
splitSequenceIntoSubstrings(obj)
}
\arguments{
\item{obj}{The EOS report object}
}
\value{
The object with left_seq and right_seq slots filled.
}
\description{
Function for splitting sequence into subsequences
This function splits sequences into left and right sequences based on cdr3Index.  Must have cdr3Index, ndnIndex, ndnLength and trimmed_sequence slots filled.  Call determineNdnValues, determineCdr3Index and trimCloneSequence functions first.
}
