% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_functions.R
\name{stability_model_selection}
\alias{stability_model_selection}
\title{Function for selecting stability models}
\usage{
stability_model_selection(MRD, Time, Sample.ID, stability.plot = TRUE,
  UL = 1.3, LL = 0.7)
}
\arguments{
\item{MRD}{The measured MRD values}

\item{Time}{Time (numeric)}

\item{Sample.ID}{Sample ID used for model selection}

\item{stability.plot}{If TRUE, returns a ggplot of the data}

\item{UL}{Upper limit of MRD (in MRD raw scale)}

\item{LL}{Lower limit of MRD (in MRD raw scale)}
}
\value{
Returns a list containing a table of the ANCOVA analyses, the selected model, the estimated shelf_life and a plot if specified.
}
\description{
Function for selecting stability models
}
