% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{methods}
\name{trimCloneSequence,EOS_Report,numeric-method}
\alias{trimCloneSequence,EOS_Report,numeric-method}
\title{trimCloneSequence}
\usage{
\S4method{trimCloneSequence}{EOS_Report,numeric}(obj, seq_edge_trim = 2)
}
\description{
trimCloneSequence
}
