% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{trimCloneSequence}
\alias{trimCloneSequence}
\title{Function for generating trimmed sequences}
\usage{
trimCloneSequence(obj, seq_edge_trim)
}
\arguments{
\item{obj}{The EOS report object}

\item{seq_edge_trim}{how many base pairs to trim on each side}
}
\value{
The object with trimmed_sequence slot filled.  "N" base pairs are also removed.
}
\description{
Function for generating trimmed sequences
}
